commission fee to float

+++ flaskr/product_mongo/basic_doc.py
@@ -275,7 +275,7 @@ class LineItem(Document):
     product = LazyReferenceField('Product')
     order = LazyReferenceField('Order')
     variant = LazyReferenceField('ProductVariant')
-    commission_fee = DecimalField(required=True, help_text="The commission fee = price*quantity*commission")
+    commission_fee = FloatField(required=True, help_text="The commission fee = price*quantity*commission")
     commission = StringField(help_text='commission of commission fee')
